pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
manage-fastapi = "^1.1.1"
fastapi = "^0.95.2"
pydantic = "^1.10.8"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.6"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.41.1"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
alembic = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
